Store input images as doubles instead of uint8_t

+++ data_io.h
@@ -10,7 +10,7 @@
 
 // Images from a to b (exclusive) are read. An image is represented as a
 // one-dimensional array organized by rows.
-uint8_t **read_images(char const *const image_fname, size_t a, size_t b);
+double **read_images(char const *const image_fname, size_t a, size_t b);
 
 uint8_t *read_labels(char const *const label_fname, size_t a, size_t b);
